Small improvements

+++ src/neural_network/layer/perceptron/activation_function/ImprovedSigmoid.hpp
@@ -5,12 +5,10 @@ namespace snn::internal
 {
     class ImprovedSigmoid : public ActivationFunction
     {
-    private :
-
+    private:
         activationFunction getType() const override { return iSigmoid; }
 
     public:
-
         float function(const float x) const override
         {
             return 1.0f / (1.0f + exp(-x)) + x * 0.05f;
@@ -18,7 +16,7 @@ namespace snn::internal
 
         float derivative(const float x) const override
         {
-            return exp(x) / pow((exp(x) + 1.0f), 2);
+            return exp(x) / powf((exp(x) + 1.0f), 2);
         }
     };
 }
